moved playSound to its own util folder & moved all 'handleInput' functions to App.js
